# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Digital Twin Fisiologico API",
    "description": "Stateless physiology analytics API. HTTP routers are thin; application services orchestrate engines under `engines/`.",
    "version": "5.1.0",
    "x-codegen": {
      "typescript_output": "frontend/src/api/generated/schema.ts",
      "regenerate": "make openapi-frontend"
    }
  },
  "paths": {
...
    "/workouts/validate": {
      "post": {
        "tags": [
          "workouts"
        ],
        "summary": "Validate workout template",
        "operationId": "workoutsValidate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Engine JSON payload — see docs/FRONTEND_DEVELOPER_GUIDE.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnginePayload",
                  "properties": {},
                  "additionalProperties": true,
                  "type": "object",
                  "title": "EnginePayload",
                  "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workouts/prescribe": {
      "post": {
        "tags": [
          "workouts"
        ],
        "summary": "Prescribe workout to athlete watts",
        "description": "Resolve percentage targets to concrete watts/HR for the athlete profile.",
        "operationId": "workoutsPrescribe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutPrescribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Materialised workout prescription",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutPrescribeResponse",
                  "properties": {
                    "status": {
                      "type": "string",
                      "title": "Status",
                      "examples": [
                        "success"
                      ]
                    },
                    "prescription": {
                      "additionalProperties": true,
                      "type": "object",
                      "title": "Prescription"
                    },
                    "athlete_profile_used": {
                      "additionalProperties": true,
                      "type": "object",
                      "title": "Athlete Profile Used"
                    }
                  },
                  "type": "object",
                  "required": [
                    "status",
                    "prescription",
                    "athlete_profile_used"
                  ],
                  "title": "WorkoutPrescribeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workouts/feasibility": {
      "post": {
        "tags": [
          "workouts"
        ],
        "summary": "Preview workout feasibility",
        "description": "W′ balance simulation before calendar assignment.",
        "operationId": "workoutsFeasibility",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutFeasibilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Engine JSON payload — see docs/FRONTEND_DEVELOPER_GUIDE.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnginePayload",
                  "properties": {},
                  "additionalProperties": true,
                  "type": "object",
                  "title": "EnginePayload",
                  "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workouts/compare": {
      "post": {
        "tags": [
          "workouts"
        ],
        "summary": "Compare assigned vs performed workout",
        "description": "Compliance score between prescribed workout and performed FIT/power stream.",
        "operationId": "workoutsCompare",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_workoutsCompare"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Engine JSON payload — see docs/FRONTEND_DEVELOPER_GUIDE.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnginePayload",
                  "properties": {},
                  "additionalProperties": true,
                  "type": "object",
                  "title": "EnginePayload",
                  "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Payload too large",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workouts/calendar/transition": {
      "post": {
        "tags": [
          "workouts"
        ],
        "summary": "Validate calendar assignment transition",
        "operationId": "workoutsCalendarTransition",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarTransitionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Engine JSON payload — see docs/FRONTEND_DEVELOPER_GUIDE.md",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnginePayload",
                  "properties": {},
                  "additionalProperties": true,
                  "type": "object",
                  "title": "EnginePayload",
                  "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "description": "Human-readable message or structured error object."
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "title": "ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_workoutsCompare": {
        "properties": {
          "workout_json": {
            "type": "string",
            "title": "Workout Json",
            "description": "Assigned workout JSON."
          },
          "athlete_profile_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Athlete Profile Json"
          },
          "tolerance_policy_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tolerance Policy Json"
          },
          "file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "power_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power Json"
          }
        },
        "type": "object",
        "required": [
          "workout_json"
        ],
        "title": "Body_workoutsCompare"
      },
      "CalendarTransitionRequest": {
        "properties": {
          "current_status": {
            "type": "string",
            "title": "Current Status",
            "description": "Current assignment status in calendar FSM."
          },
          "desired_status": {
            "type": "string",
            "title": "Desired Status",
            "description": "Requested next status."
          }
        },
        "type": "object",
        "required": [
          "current_status",
          "desired_status"
        ],
        "title": "CalendarTransitionRequest"
      },
...
      "EnginePayload": {
        "properties": {},
        "additionalProperties": true,
        "type": "object",
        "title": "EnginePayload",
        "description": "Open-ended JSON returned by physiology engines.\n\nField semantics, tiers and confidence rules are documented in\n``docs/FRONTEND_DEVELOPER_GUIDE.md``. Treat unknown keys as forward-compatible."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkoutFeasibilityRequest": {
        "properties": {
          "workout": {
            "additionalProperties": true,
            "type": "object",
            "title": "Workout"
          },
          "athlete_profile": {
            "additionalProperties": true,
            "type": "object",
            "title": "Athlete Profile"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context",
            "description": "Optional readiness/fatigue context for W′ simulation."
          }
        },
        "type": "object",
        "required": [
          "workout"
        ],
        "title": "WorkoutFeasibilityRequest"
      },
      "WorkoutPrescribeRequest": {
        "properties": {
          "workout": {
            "additionalProperties": true,
            "type": "object",
            "title": "Workout"
          },
          "athlete_profile": {
            "additionalProperties": true,
            "type": "object",
            "title": "Athlete Profile",
            "description": "Athlete CP/FTP/weight used to resolve percentage targets."
          }
        },
        "type": "object",
        "required": [
          "workout"
        ],
        "title": "WorkoutPrescribeRequest"
      },
      "WorkoutPrescribeResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "examples": [
              "success"
            ]
          },
          "prescription": {
            "additionalProperties": true,
            "type": "object",
            "title": "Prescription"
          },
          "athlete_profile_used": {
            "additionalProperties": true,
            "type": "object",
            "title": "Athlete Profile Used"
          }
        },
        "type": "object",
        "required": [
          "status",
          "prescription",
          "athlete_profile_used"
        ],
        "title": "WorkoutPrescribeResponse"
      },
      "WorkoutValidateRequest": {
        "properties": {
          "workout": {
            "additionalProperties": true,
            "type": "object",
            "title": "Workout",
            "description": "Workout template or coach draft with steps."
          }
        },
        "type": "object",
        "required": [
          "workout"
        ],
        "title": "WorkoutValidateRequest"
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "{baseUrl}",
      "description": "Configurable deployment",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:8000",
          "description": "Set via VITE_API_BASE_URL in the frontend"
        }
      }
    }
  ],
...
}